requirements.txt
celery==4.0.2
django==1.11.8
django-cors-headers==2.1.0
django-celery-results==1.0.1
djangorestframework==3.6.3
flower==0.9.2
mysqlclient==1.3.10
python-can==1.5.2
requests
transitions==0.5.3